+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2023-10-15
+### Fixed
+- FFI Arc Course Calculations function name.
 
 ## [1.1.0] - 2023-10-15
 ### Added
@@ -193,8 +196,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Math Tools
   - Conversion Functions
   - Legendre Spherical Harmonics Functions (For Magnetic Calculations)
-  
 
+[1.1.1]: https://github.com/997R8V10/aviation-calc-util/compare/v1.1.1...v1.1.1
 [1.1.0]: https://github.com/997R8V10/aviation-calc-util/compare/v1.0.20...v1.1.0
 [1.0.20]: https://github.com/997R8V10/aviation-calc-util/compare/v1.0.16...v1.0.20
 [1.0.16]: https://github.com/997R8V10/aviation-calc-util/compare/v1.0.12...v1.0.16